requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.4.0
openai==1.10.0
boto3==1.34.29
django-storages==1.14.2
whitenoise==6.6.0
dj-database-url==2.1.0
django-filter==23.5
requests==2.31.0

